WIP: had to rename. Having trouble with .decode in Node
