+++ jest.config.js
@@ -1,4 +1,5 @@
 module.exports = {
+  "testEnvironment": "jsdom",
   "coverageDirectory": "coverage",
   "coverageReporters": ["text", "html"],
   "roots": [
