+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './reactChromakeyedImage';
+export * from './ChromakeyedImage';
 
 export * from './colorStringUtils';
 export * from './pixelUtils';
